Struct nanorand::buffer::BufferedRng

pub struct BufferedRng<InternalGenerator: Rng<OUTPUT>, const OUTPUT: usize> { /* private fields */ }

A buffered wrapper for any Rng implementation. It will keep unused bytes from the last call to Rng::rand, and use them for subsequent randomness if needed, rather than throwing them away.

use nanorand::{Rng, BufferedRng, WyRand};

let mut thingy = [0u8; 5];
let mut rng = BufferedRng::new(WyRand::new());
rng.fill(&mut thingy);
// As WyRand generates 8 bytes of output, and our target is only 5 bytes,
// 3 bytes will remain in the buffer.
assert_eq!(rng.buffered(), 3);

Implementations

impl<InternalGenerator: Rng<OUTPUT>, const OUTPUT: usize> BufferedRng<InternalGenerator, OUTPUT>

pub fn new(rng: InternalGenerator) -> Self

Wraps a Rng InternalGenerator in a BufferedRng instance.

pub fn into_inner(self) -> InternalGenerator

Returns the internal RNG, dropping the buffer.

pub fn buffered(&self) -> usize

Returns how many unused bytes are currently buffered.