Trait nanorand::gen::RandomGen

pub trait RandomGen<Generator: Rng<OUTPUT>, const OUTPUT: usize> {
    fn random(rng: &mut Generator) -> Self;
}

A trait used for generating a random object with an RNG,

Required Methods

fn random(rng: &mut Generator) -> Self

Return a random instance of the implementing type, from the specified RNG instance.