Trait nanorand::rand::Rng

pub trait Rng<const OUTPUT: usize>: Clone {
    // Required method
    fn rand(&mut self) -> [u8; OUTPUT];

    // Provided methods
    fn generate<Generated>(&mut self) -> Generated
       where Generated: RandomGen<Self, OUTPUT> { ... }
    fn fill_bytes<Bytes>(&mut self, buffer: Bytes)
       where Bytes: AsMut<[u8]> { ... }
    fn fill<Contents, Array>(&mut self, target: Array)
       where Contents: RandomGen<Self, OUTPUT>,
             Array: AsMut<[Contents]> { ... }
    fn generate_range<Number, Bounds>(&mut self, range: Bounds) -> Number
       where Number: RandomRange<Self, OUTPUT>,
             Bounds: RangeBounds<Number> { ... }
    fn shuffle<Contents, Array>(&mut self, target: Array)
       where Array: AsMut<[Contents]> { ... }
}

A trait that represents a random number generator.

Required Methods

fn rand(&mut self) -> [u8; OUTPUT]

Generates a random sequence of bytes, seeding from the internal state.

Provided Methods

fn generate<Generated>(&mut self) -> Generatedwhere Generated: RandomGen<Self, OUTPUT>,

Generates a random of the specified type, seeding from the internal state.

fn fill_bytes<Bytes>(&mut self, buffer: Bytes)where Bytes: AsMut<[u8]>,

Fill an array of bytes with randomness.

fn fill<Contents, Array>(&mut self, target: Array)where Contents: RandomGen<Self, OUTPUT>, Array: AsMut<[Contents]>,

Fill an array with the specified type.

fn generate_range<Number, Bounds>(&mut self, range: Bounds) -> Numberwhere Number: RandomRange<Self, OUTPUT>, Bounds: RangeBounds<Number>,

Generates a random of the specified type, seeding from the internal state.

fn shuffle<Contents, Array>(&mut self, target: Array)where Array: AsMut<[Contents]>,

Shuffle a slice, using the RNG.